fixed satya merge
